requirements.txt
streamlit==1.37.0
tensorflow
opencv-python-headless
pillow
numpy